Visual studio bug fixed, remove pragma warning suppression
also make pointer ownership explicit, remove more pragma warning


formatting

+++ Source/Misc.cpp
@@ -137,22 +137,23 @@ namespace {
 
 std::wstring rsj::Utf8ToWide(std::string_view input)
 { // add terminating null
-   const auto buffersize =
-       MultiByteToWideCharErrorChecked(CP_UTF8, 0, input.data(), input.size(), nullptr, 0) + 1;
+   const auto buffersize = MultiByteToWideCharErrorChecked(CP_UTF8, 0, input.data(),
+                               gsl::narrow_cast<int>(input.size()), nullptr, 0)
+                           + 1;
    std::vector<wchar_t> buffer(buffersize, 0); // all zero
-   MultiByteToWideCharErrorChecked(
-       CP_UTF8, 0, input.data(), input.size(), buffer.data(), gsl::narrow_cast<int>(buffer.size()));
+   MultiByteToWideCharErrorChecked(CP_UTF8, 0, input.data(), gsl::narrow_cast<int>(input.size()),
+       buffer.data(), gsl::narrow_cast<int>(buffer.size()));
    return buffer.data();
 }
 
 std::string rsj::WideToUtf8(std::wstring_view wstr)
 { // add terminating null
-   const auto buffersize = WideCharToMultiByteErrorChecked(
-                               CP_UTF8, 0, wstr.data(), wstr.size(), nullptr, 0, nullptr, nullptr)
+   const auto buffersize = WideCharToMultiByteErrorChecked(CP_UTF8, 0, wstr.data(),
+                               gsl::narrow_cast<int>(wstr.size()), nullptr, 0, nullptr, nullptr)
                            + 1;
    std::vector<char> buffer(buffersize, 0);
-   WideCharToMultiByteErrorChecked(CP_UTF8, 0, wstr.data(), wstr.size(), buffer.data(),
-       gsl::narrow_cast<int>(buffer.size()), nullptr, nullptr);
+   WideCharToMultiByteErrorChecked(CP_UTF8, 0, wstr.data(), gsl::narrow_cast<int>(wstr.size()),
+       buffer.data(), gsl::narrow_cast<int>(buffer.size()), nullptr, nullptr);
    return buffer.data();
 }
 #else
